re-merge constant.go into one source file
